scheme_env/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-scheme_env"
version = "0.1.0"
description = "Indian Government Welfare Officer RL Environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "scheme_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["scheme_env", "scheme_env.server"]
package-dir = { "scheme_env" = ".", "scheme_env.server" = "server" }